Add proper error handling to the shell library's control port interaction.
It used to mask actual problems with talking to the control port by
returning an empty answer, which sometimes is expected (e.g. it can be
for`tor_control_getconf Bridge`).

Now we detect and return errors in all cases I could find so callers
that use `set -e` or checks the return code benefits.

In particular, for #18293 this will treat the situation in the
automated test suite where `tor_control_getconf DisableNetwork`
receives an empty answer, and properly treat that as a failure; the
empty answer we otherwise would receive would lead to a logic error,
so best to abort before that.