TCA: Only save known-working Tor configuration to tca.conf
So far the "tor" section of tca.conf was write only so it did not matter
whether we saved potentially broken configuration there. But we'll
soon start loading this information from tca.conf, so better ensure
we only save info that works.

Note: this is not ideal because with this change, upon connection failure,
previously working configuration is now deleted from tca.conf.
Currently this does not matter since we never load that configuration,
and for #5461 I'll move the "ui" bits to another file, after which
we can save them unconditionally there and leave tca.conf unchanged
in the successful_connect=False situation.

So really this commit is only useful to give myself a baseline
that makes enough sense to me that I can then modify it.

refs #5461