Test suite: disable buggy test scenario
This essentially reverts ea85291ad25e148b16f2568d079720c078c37f47, which
I merged earlier today, because it introduced a regression: it assumes that in
the current state of things (esp. wrt. #18470), in "Scenario TCA can reconnect
after a connection failure" we will always end up using the default bridges,
which is incorrect. At least in some cases we are not going to use bridges,
because tor won't have bootstrapped yet by the time we restart Tor Connection
and apply UseBridges 0. I've seen this happen on a local test suite run.

And while I'm at it, I'm explaining what the current behavior is,
and what the desired, future behavior will be.

refs #18493