TCA: remove incorrect signal handling
So far this problem was not visible because I think there was no way to trigger
this signal, but now that we have keyboard accelerators for these radio buttons,
the second time one uses the "a" or "h" accelerator, it triggers the signal and
thus connects to Tor.

IMO we should never call cb_step_hide_btn_connect_clicked unless
the user explicitly clicked the "Connect to Tor" button,
or did $something that clearly expresses the intent to do so.

Activating that button when one of the radio buttons is activated
is very uncommon behavior in the GTK world and can be surprising
to users, which is no good in our consent screen.