Test suite: prepare sandbox test for when #18470 is fixed
Given the current implementation of the test about tor seccomp sandbox, there's
no simple way to make "Tor is ready" pass reliably in the "TCA can reconnect
after a connection failure" scenario.

But once #18470 is fixed, this should be straightforward, so here
I'm adding the code we'll need once we're there.