Avoid spamming new python interpreters while waiting for Tor to bootstrap.
Now that the tor.sh shell library is based on python-stem, each call
spawns a python interpreter, which is pretty heavy. Especially when
spamming tor_is_working() like when waiting for Tor to bootstrap. So
we provide tor_wait_until_bootstrapped() which does all of it in once
instance of python interpreter.