Install squashfs-tools from bullseye.
In sid, where we used to install it from for obsolete reasons, it was
just bumped to version 1:4.5-2 which adds a dependency on libzstd1 >=
1.4.0, which isn't in buster.

However, sid's libzstd1 (currently 1.4.8+dfsg-2.1, which also is in
bullseye) installs just fine in current Tails, so we could try that
route instead *if* we want squashfs-tools to keep rolling with sid. At
the moment I see no reason for that.