Explain better how to type a bridge (#18597)
During the usability tests in August, people were very confused about
the procedure to follow here.

The phrase "obfs4 bridges start with the word 'obfs4'" made some
participants believe that they had to start typing 'obfs4' and wondered
about what else to type.

I'm clarifying the procedure with the phone. I used "type" on purpose in
the phone example to be more explicit to the vast majority of people who
will indeed type it on their keyboard.

People were concerned about the security aspects of asking bridges over
email. Giving it as an example in the UI could help people feel more
reassured. I'm not sure that we should be more verbose in the UI itself
and the doc already explains it better.

Nowadays bridges@torproject.org and (bridges.torproject.org) only return
obfs4 by default, so we don't need to warn about what type of bridge to
use. We'll get back to it if bridges@torproject.org or
bridges.torproject.org ever returns by default a type of bridge that is
not supported in Tails.