Upgrader: use apparent size to compute space needed to install an IUK
gensquashfs uses sparse files at least for files that contain only zeroes.
As a consequence, "du" returns 0 as the non-apparent size of a directory which
contains only sparse files, which makes the IUK test suite fail.

I'm not sure if this can happen in a real IUK: hopefully we don't ship files
filled with zeroes in there. But let's play it safe and take into account the
space occupied by these zeroes.

Accordingly, in the test suite, allow some fuzziness in the apparent size
we get.

refs #18577