Drop VirtualBox.
Installing VirtualBox from Sid (since that's the only option now that
it's not been in Debian Stable since Stretch) results in this on
feature/bullseye (and something similar on devel):

The following packages have unmet dependencies:
 virtualbox-guest-utils : Depends: libc6 (>= 2.32) but 2.31-13+deb11u2 is to be installed
 virtualbox-guest-x11 : Depends: libc6 (>= 2.32) but 2.31-13+deb11u2 is to be installed

Last we discussed about this [0] we "decided to do it on a best effort
basis, i.e. to install it from sid when it's possible, and to drop it
when it's not possible anymore".

Seems like the time has come to drop proper Virtualbox.

[0] commit 4e378e52b17b6b4c737bd817c61673568203743d