Fix FTBFS by dropping VirtualBox for now (Closes #18643)
Installing the VirtualBox guest support from Sid (since that's the
only option now that it's not been in Debian Stable since Stretch)
now results in this:

The following packages have unmet dependencies:
 virtualbox-guest-utils : Depends: libc6 (>= 2.32) but 2.28-10 is to be installed
 virtualbox-guest-x11 : Depends: libc6 (>= 2.32) but 2.28-10 is to be installed

Last we discussed this [0] we "decided to do it on a best effort
basis, i.e. to install it from sid when it's possible, and to drop it
when it's not possible anymore". So it might be time for that!
However, on #18643 we are considering to install them from Debian Fast
Track, so proper VirtualBox guest support might still return.

[0] commit 4e378e52b17b6b4c737bd817c61673568203743d