Test suite: give GNOME Overview more time to process our search
I've seen
a case (https://gitlab.tails.boum.org/tails/tails/-/merge_requests/647#note_179280)
in which this code tried to type "c", then "ommandline", to start
GNOME Terminal.

But instead, "Tor Connection" was selected and started, because GNOME
did not have time to get all the keyboard input and process it,
and it was a good match for the "co" search string.

Let's give GNOME more time to process our keyboard input and to generate
search results.