Test suite: stop tor at scenario end to prevent leaks
I've seen this at the beginning of scenarios, where 10.2.1.1 is the
host and 10.2.1.94 is the system under testing:

  Time        Source      Dest        Proto Len   Info
  3.028186    10.2.1.1    10.2.1.94   TCP	  66    9940 → 45256 [FIN, ACK] Seq=1 Ack=1 Win=501 Len=0 TSval=1801081331 TSecr=3040065954
  3.028214    10.2.1.1    10.2.1.94   TCP	  1514  9940 → 45264 [ACK] Seq=1 Ack=1 Win=501 Len=1448 TSval=1801083187 TSecr=3040069093
  3.328009    10.2.1.1    10.2.1.94   TCP	  66    [TCP Retransmission] 9940 → 45256 [FIN, ACK] Seq=1 Ack=1 Win=501 Len=0 TSval=1801084659 TSecr=3040065954
  3.328360    10.2.1.94   10.2.1.1    TCP	  54    45256 → 9940 [RST] Seq=1 Win=0 Len=0

This breaks most of tor_bridges.feature: when a previous scenario
bleed a packet into the next, that TCP RST might be sent to a host the
scenario is not expecting. Gracefully stopping tor prevents this.