Vagrant build box: have hostname resolve to 127.0.0.1
Quoting anonym:

"The reason is that I sometimes experience a bug with QEMU where my builder will
lose network connectivity. I can save the build by replugging its (virtual)
ethernet cable and restarting the networking service, but due to $HOSTNAME
resolving to the LAN IP address by default (due to NetworkManager, I think)
I have to wait a long time for the DNS resolution to timeout. BTW, this used to
be the case when we used vmdebootstrap, so it kinda is a regression. :)"