Select iptables-legacy as our iptables implementation
Buster defaults to iptables-nft, which ferm doesn't support so it
explicitly looks for the iptables-legacy. We don't want to mix -legacy
and -nft so let's stick with -legacy for all for now. In the future
we'll probably drop ferm and switch to -nft (which has most of ferm's
improvements over -legacy).

Closes tails/tails#18706