Always display the 1st screen if there's any error (#18720)
If the user has a Persistent Storage and an administration password, we
can't know if they have a backup Tails already or not. Asking them to
plug it in if they don't have one will be very confusing.

The 1st screen is most important as orientation for people who might
give a shot at the application without having a backup Tails already.

For people how already have a backup Tails and might plug it in on
"Retry", I think it's good to always display them to 2nd confirmation
screen. It gives a confirmation that the initial error has been solved
and reminds that all backup data will be replaced.