Test suite: detect disks being plugged using udisks
The 'I plug $BUS drive ...' step waits until the drive is detected,
and but let's make that be when udisks detects it, since some things
we do after plugging a device depends on that, not just that the block
device was created.

And let's make use of this in the recently added scenario where we had
to wait for udisks after plugging the drive.