Welcome Screen: finally, drop unneeded, problematic code
So something has changed in Bullseye so the UserManager isn't ready as
early as before. However, it is ready when we actually need it to,
i.e. when calling the final LanguageSetting.apply_language() when
logging in.

FWIW, when the Welcome Screen initializes (specifically
LanguageSettingUI) there is another call to apply_language() which
causes us to log:

  language.py:194 apply_language: AccountsManager not ready

but since we at that time do not case that the Live user got the
correct language set (we only care about the Welcome Screen's GUI) it
can be safely ignored.

Closes tails/tails#18693