gdm-tails.json: remove panelStyle
As segfault noticed on #12551 a few years ago, GNOME uses login-screen,
and loginScreen is deprecated.

Our previous value ("loginScreen") seems to be a no-op as it's nowhere in the
GNOME Shell source code.

The new value ("login-screen") makes the text of the top bar white on light
blue, which is not readable enough.

refs #16951