Drop obsolete patch
I can't find where the patched script is used and I can't figure out why the
operations it does are necessary.

With a Bullseye image, I've `rgrep --text live-medium-cache` (the name of the patched script) in:

- the mounted `filesystem.squashfs`
- the unpacked initramfs

⇒ no occurrence besides the obvious one that merely indicates the script is installed: `var/lib/dpkg/info/live-tools.list:/bin/live-medium-cache`

Same with 4.27.

So either the name of that script is generated programmatically by its caller, or it's never used.

There's no occurrence of "cache" in the versions of `live-boot`, `live-tools`, and `live-config` installed in our Bullseye images. Same in the versions installed in 4.27.

Actually, it seems the code this patch modifies has never been used since a long
time: it's been moved from `live-config` in version 4.0~a24 to `live-tools`,
then factored out in
https://salsa.debian.org/live-team/live-tools/-/commit/37dd2f4e7aa76a230b7a42ac6691327150562c07,
which moved code to a new script… without ever calling that new script.
We followed suit in ae54b6245602ebeb3c51191ec7dbfef97960d629 to patch that new
script in Tails 2.0. And since more than 7 years we've kept updating our patch.

refs #18691, #6228