tails-get-network-time: emulate NetworkManager's behavior more closely
Like NetworkManager, we abort the connection as soon as
we have enough information to draw conclusions.

With one difference though: even if we get "X-NetworkManager-Status: online"
early, we can't abort the connection until we've received the Date header.
In theory this is fingerprintable:

 - not exploitable by a passive adversary as long as we use a URL whose server does not return X-NetworkManager-Status: online: we're good with current Fedora server
 - exploitable by an active adversary: inject this HTTP header in the legit response and see us not abort the connection, while NM would.

But I doubt this is a realistic attack: I suspect the first batch of data sent by the HTTP server is large enough to include all headers.