Install obfs4proxy from the Tor Browser tarball
The main goal here is to ship an up-to-date obfs4proxy, since the version in
Debian is vastly outdated and has a different network fingerprint
than the most commonly used by Tor clients, which is the one shipped
in Tor Browser.

In particular, we want to fix the Elligator2 bug (tails#18800). This will be
achieved automatically, thanks to this commit, once we upgrade to Tor Browser
which has upgraded obfs4proxy:
https://gitlab.torproject.org/tpo/applications/tor-browser-build/-/issues/40416

And while I'm at it, I'm installing all Pluggable Transports shipped in the Tor
Browser tarball. Currently this means we also install the snowflake-client
binary, which is useless right now but should save us some time when we work on
tails#5494. And more generally, this approach should make it slightly easier for
us to pick up support for PTs newly added in Tor Browser.

refs #18800, #5494