UnsafeBrowser tests don't need Tor; scenarios--
No tests about UnsafeBrowser was using Tor; also, we are suggesting the
use of UnsafeBrowser _only_ when Tor is not working and they need to
setup a captive portal, so now our tests align better with what users
do.

The no-network-logged-in-unsafe-browser scenario was only used in one
case. When we need it, we can just unplug the network manually.