Test suite: refactor and fix re/dis-connecting to/from a network interface via the system menu
For the life of me I can't get this to work on Bullseye: clicking on "Wired
Connected" with Dogtail closes the System menu, instead of opening the network
interface configuration sub-menu. Same if I click with xdotool. I verified in
interactive debugging mode that the coordinates are correct and the .point
method highlight the right screen area.

While this works all fine when I take over TailsToaster with virt-viewer and
test this manually. Go figure.

@anonym noticed the same in b805f1f282f08de15a36fee26bcd9175b6dd8235, and
adjusted another scenario to cope with it. So I'm refactoring anonym's code into
a method and using it in the 2 places where we interact with the system menu in
our test suite.