Release process: don't wait for something that can take up to 1 hour
This Weblate integration process runs only once an hour, so it's a vastly
over-cautious indicator that the website has been fully refreshed.
Let's not add so much delay in the release process here.
Instead, use simpler heuristics, that may be slightly incorrect,
but should be good enough for our purposes.