Test suite: allow the pattern to cover slightly less of the write cache
I've seen a case when it covered "only" 96%. IMO that's good enough.
I suppose more stuff is happening on Bullseye which can occasionally
make the kernel reclaim that some of write cache. Or, the kernel
already started to write this data out to disk, and thus removed
it from its write cache.