Additional Software: synchronize APT data when needed directly from t-p-s
For some reason I could not understand yet,
tails-synchronize-data-to-new-persistent-volume-on-shutdown.service
does not always manage to fully do its job.

So let's at least make things work when the user clicks "Restart"
in t-p-s — which is the most obvious path for them — by synchronizing
data there *before* we trigger the reboot.

refs #18839