Install Tails' fork of live-build from a submodule.
Previously, live-build was installed from custom-built Debian
packages.  This patch instead adds the custom fork as a git submodule,
cutting out the packaging from development and testing.

@intrigeri notes:

> After second thought, maybe it's a good time to stop using a (custom
> fork of an ancient) Debian package for live-build: this makes it
> difficult to prepare and correctly test such contributions (you
> basically have to build & upload a package to test any change), and
> painful to review/test/accept them. I already noticed that when I
> was working on squashfs-tools-ng and zstd compression myself
> recently. I'm under the impression that historically, this has
> negatively impacted our motivation to work on anything that required
> patching live-build.
>
> So, we could:
>
> 1. Add our live-build repo as a Git submodule of tails/tails, in some
>    relevant place under vagrant/
>
> 2. In the Vagrant box build process, somehow copy the relevant
>    live-build files to the VM filesystem; this may boil down to running
>    DESTDIR=something make install

These steps are done.

> 3. In the code that determines which Vagrant box shall be used (and if
>    necessary, first built): take into account the status of the
>    live-build submodule, if not done for free thanks to step 1

See box_name in vagrant/lib/tails_build_settings.rb.  That comes down
to git log -1 ./vagrant, so it will not pick up changes to
./submodules/live-build.

However, because we install live-build in build-tails, live-build is
not actually part of the basebox image, therefore the concern is
addressed.

Furthermore, this change drops the 'live-build' line from
/etc/amnesia/version.  The information which version of live-build is
used is now implicitly included in Tails' git commit id because
live-build is now a submodule of that repository.