Remove unused IncorrectOwnerException and IncorrectOwnerError
Apparently I used to raise those when activating symlinks and the
destination directory was owned by a different user. The only test which
checked for that created a symlink to a directory owned by a different
user - that case is now handled by using the nosymfollow mountpoint, so
I hope that there is really no need anymore for this error.