Adjust unwanted packages removal for Buster
When using tagged APT snapshots, attempting to remove gcc-8 triggers a FTBFS,
because APT has never heard of gcc-8.

Let's instead use regexps: they will only match packages that APT knows about.
And, as a bonus, we won't have to update it every time we switch to
a new Debian.