Test suite: support Bookworm host system's improved UEFI graphics
Until Bullseye, our background picture was deformed in UEFI mode: the aspect
ratio was not preserved. That's fixed on current sid, so will presumably be
fixed in Bookworm. I suppose this came with an update of the OVMF firmware or
of QEMU.

Similarly, on Bookworm the whole GRUB screen does not fit into what we're
capturing, so we don't see the end of the GRUB command line, and there as well
we need a different expected image.