Use Bullseye debootstrap configuration
In the live-build code, this translates into
/usr/share/debootstrap/scripts/${LB_BOOTSTRAP_CONFIG}. Currently both "buster"
and "bullseye" in there are symlinks to "sid", so this change should be a no-op.

However, this may change in the future. For example, if this changes for
Bookworm, and if we keep hard-coding "buster" in auto/config, then we'll likely
miss it when we look for instances of "bullseye" that we should replace with
"bookworm", and we'll be stuck to using the configuration meant for Buster,
which could cause all sorts of problems.