Test suite: give the XMPP server some time to create the room
Recently I've seen numerous cases where the "Create new room" dialog is
displayed pretty quickly, but it is initially empty, and only after a few
seconds the "Accept Defaults" button appears. Without this change,
when this happens we would fail because we assumed that button would
be on screen as soon as the dialog appeared.