Test suite: Fix clock upper bound calculation
I've seen this fail:

    Given I have started Tails from DVD without network and logged in
    When I bump the hardware clock's time with "-15 days"
    And I warm reboot the computer
    And the computer reboots Tails
    Then the hardware clock is still off by "-15 days"

With that error:

    The host's hwclock should be approximately '2022-04-27 02:24:44 +0000' but is actually '2022-04-27 02:24:59 +0000'.
    <false> is not true. (Test::Unit::AssertionFailedError)
    ./features/step_definitions/time_syncing.rb:116:in `/^the hardware clock is still off by "([^"]+)"$/'
    features/time_syncing.feature:46:in `Then the hardware clock is still off by "-15 days"'

Given expected_time_lower_bound == expected - max_time_drift,
then expected_time_upper_bound == expected_time_lower_bound + max_time_drift
                               == expected,
which means we did not accept the clock to be in the future at all.

Let's accept that it can be up to 1 second in the future, which happened
in the test suite run I quoted above.