Test suite: enable debug logging for Screen#wait
I see a number of weird failures where an image is not found on the screen,
while I would expect it to be: calling opencv_match_template.py with the
same arguments our test suite does in the context of those failures
does find the image on the screen.

So I suspect something down the call chain is slower than it used to be (taking
the screenshot? OpenCV?), which could make the try_for timeout before the image
has actually appeared.

This debug logging will give us timing information that'll be useful to tell
whether this is actually the case.