Test suite: make copying'n'pasting into a Terminal more robust
Since !821 I occasionally see failures where we paste an empty string into
a Terminal, just after having put the desired, non-empty, string into
the clipboard.

The typical usage pattern looks like this:

00:15:00.620404108: Remote shell: calling as amnesia: echo -n 'pkexec touch /root/pkexec-test' | xsel --input --clipboard
00:15:00.829120193: Remote shell: call returned: [0, "", ""]
00:15:00.829343562: Keyboard: pressing: ctrl+shift+v
00:15:01.892986705: Keyboard: pressing: Return

I did not notice this elsewhere than in Terminal but I did not check
thoroughly.

So it seems that somehow, the contents of the clipboard, as seen by GNOME
Terminal, is lagging behind what we would like it to be.

Let's make this more robust in 2 ways:

 - When pasting into a Terminal, let's wait a bit before pasting.
 - In all cases, ensure that set_clipboard has effectively
   produced the desired effect.