Test suite: avoid missed clicks retry when opening the calendar & notifications menu
We often miss clicks when interacting with the GNOME top bar, so let's use the
keyboard shortcut instead. This should be sufficient, but just in case, let's
integrate this into the retry_action loop… which incidentally makes
its recovery_proc meaningful: if we don't try opening that menu between
attempts, pressing Escape is not going to help.