Test suite: also set the guest's time when connected to the LAN but not to Tor
At some point, we've introduced a conditional sub-branch that does not
sync the time from host to guest, which can cause a variety of problems.

We used to sync the time from host to guest in *every* case,
before we restarted Tor, but it went away in a buggy merge conflict
resolution: 81fa0cab2e7475e900ce26c82e575396d6fd4c72.