Test suite: deal with Wayland vs Dogtail issue.
Under Wayland the title bar and its elements are exposed to Dogtail,
unlike under Xorg, and it contains one label for the window title (and
one invisible label of unkown use) so we can no longer assume that
there is only one label for e.g. a zenity dialog that only has one
label under Xorg.