Don't try (and fail) to refresh features' IsActive when unlocking
Refreshing IsActive would run on_activated(), which failed because at that point
the feature was not mounted yet. We'll refresh this property later on, once we
run Activate.

This fixes this bug: Unlocking a newly created Persistent Storage fails after
reboot (both in the Welcome Screen and in the amnesia GNOME session). How to
reproduce:

 - Either start Tails, create Persistent Storage with all features enabled
   except dotfiles, reboot, type passphrase in the Welcome Screen, click Unlock.
 - Or do the same, but instead of unlocking in the Welcome Screen, log in then
   try to unlock with Persistent Storage inside amnesia's GNOME session.

At this point I'm not familiar enough with this code base to be confident
I'm doing the right thing. It could be that this commit breaks other
stuff, or that there's a much more elegant way to fix this bug.
But at least this will allow me to update our test suite, which in turn
should help spot regressions I'm introducing here :)