Revert incomplete split of the Welcome Screen persistence feature
Before this, one could not login when the administration password was
persistent: `/var/lib/gdm3/settings/persistent/tails.password` was owned by
`root:root` with mode `600` so the Greeter could not read it.

This lead me to notice that only the admin password was supported on this
branch. Other Welcome Screen settings could not be made persistent anymore:
Unsafe Browser, MAC spoofing.

Additionally, the admin password is stored in a different place than
previously (`password` vs. `greeter-settings/tails.password`). This is not going
to play well with a pre-existing persistent `greeter-settings` directory that
remains used concurrently.

In the new design, these Welcome Screen settings are indeed meant to be split
into individual persistence features. IMO, for the first iteration we should
revert to a single Welcome Screen feature. We can figure out later how to do the
split, migrate configuration & data, and possibly get rid of a related
hack (#19062): there's no real benefit in doing this in the first iteration of
the Python rewrite. I'd rather focus on a smaller scope that we have a chance to
release in due time.

And anyway, this bug blocked making progress on #19001.