Test suite: run the persistence storage behave and config file unit tests
They're broken there as well at the moment (in a different manner than on GitLab
CI), but committing this will make it easier to figure out what's going on.
In particular, I find it concerning that the behave tests related to symlinks
fail inside Tails, while they work on my sid system: are we missing something in
Tails to yield the intended security properties?