Fix tps behave tests when a dedicated filesystem is mounted on /tmp
Tails has a tmpfs mounted on /tmp (and this could be the case on a developer's
machine too). This broke the behave tests, because with a non-recursive
bind-mount, context.tmpdir was not visible inside NOSYMFOLLOW_MOUNTPOINT.

Another way to fix it would be to use a rbind mount (MS_BIND + MS_REC), which
makes most of the tests pass, but then the @symlink_attack tests would fail
(as in, the mount operation that should fail succeeds) because we're mounting
somewhere under /tmp that lacks nosymfollow.