Additional Software: add more accurate wait loop
In the previous implementation, we would wait for the user to close
tails-persistence-setup, which indicates the new Persistent Storage
had been configured and we could start using it. Nowadays we can
wait for the exact pre-condition we need to proceed with the next steps:
the Additional Software persistent storage feature is active.

This is mostly a no-op for now since we're still waiting for tps to exit
(which may never happen with the new implementation), so also adding a TODO
about it.