Test suite: fix backup scenario
udisks cannot unlock 2 devices with the same UUID, and it has already unlocked
1 (the production Persistent Storage).

This problem does not happen with the old Persistent Storage implementation,
because there, live-persist unlocks the production persistent storage with
cryptsetup, which creates symlinks with different names than udisks
(e.g.
/dev/disk/by-id/dm-uuid-CRYPT-LUKS1-b694add6e2654a148b043a4103221b57-TailsData_unlocked
as opposed to udisks's
/dev/disk/by-id/dm-uuid-CRYPT-LUKS1-ffcba176fe024609914fac0ba732e378-luks-ffcba176-fe02-4609-914f-ac0ba732e378)
so there was no conflict.