Wayland vs export_gnome_env(): drop variables not dumped into /run/gnome-shell-environment/environ
Under Wayland these environment variables are not set for the
gnome-shell process. So far I have not seen failures from users of
export_gnome_env(), but if we want to fix it seems one way to do it is
to make gnome-shell-save-environment wait until XWayland is running
and then dump the environment from that process instead.