Drop incomplete attempt to support unlocking Persistent Storage after login
The design does not include this feature, but the actual implementation did.

In the current state, it's not going to work well for some features, because the
corresponding software has already started and won't re-read its configuration:
Network connections, Printers, Tor Connection, Additional Software.

Let's come back to this later if necessary.