Don't allow deleting an unlocked Persistent Storage
This feature would be nice, but it is not required in the design.

The check for conflicting apps is somewhat buggy: the modal dialog that asks me
to close the conflicting app, with a single "Cancel" button, never vanishes even
after I have closed the conflicting app; however, apparently the backend gets it
right, and the deletion operation resumes and completes successfully. But to
close the TPS GUI, I have to first click "Cancel" button (which actually does
not cancel anything), which is confusing.

We can re-enable this feature once we make it work better
(keeping in mind that for some features, even after having deleted the
Persistent Storage, the old persistent data will still be used, because the
corresponding software has already read its configuration from there).