TCA: drop useless code
Without this code, the title is correct in the window titlebar,
in the taskbar (bottom of screen), and in the GNOME top bar.
So I don't see what this code is useful for. It seems to me
that Gtk.Application and Gtk.ApplicationWindow do the right
thing already.

Closes #18610

Note: be it with or without this code, the app startup notification is somewhat
broken: the mouse cursor remains a spinner for a few seconds after the window
has been shown. So this code did not help here.