GitLab CI: run the Bandit security oriented static analyzer for Python
For now let's configure it (`-ll`) to only report only issues of severity medium
or higher: let's at least ensure we don't regress for these.

To drop this and also report issues of low severity, we would need to audit 23
pieces of code and add "nosec" comments for those that are safe.