Unsafe Browser: enable ibus and accessibility!
This is a straightforward usage of the tails-{ibus,a11y}-proxy-netns
service except it normally sets StopWhenUnneeded, which makes the
proxies instantly quit, so we override this option. It is unclear to
me why this is not the case for Tor Browser and OnionCircuits, whose
proxies are running all the time since boot, and that keep running
despite not being used.

refs #9767, #7502