Upgrader: when a manual upgrade is needed, point to the news & manual upgrade doc without specifying a version
At some point in the N+1.x life cycle, the Release Manager stops updating UDFs
for N.x. So far, the Upgrader incorrectly assumed these UDFs were always
up-to-date, and thus would keep advertising manual upgrades to an older,
obsolete N+1.x release — which is neither desirable nor possible.

Let's acknowledge that UDFs for older versions will eventually become outdated.
And thus, when a manual upgrade is advertized, and no automatic upgrade is,
let's just point to the /news and manual upgrade documentation on our website.

Of course, it would be better to:

 - Either, keep *all* old UDFs up-to-date. This would require more work
   in the release process area.

 - Or, retrieve information about the latest release
   (https://tails.boum.org/inc/stable_amd64_version/) at runtime.
   This would require more coding work than this commit.