tails-additional-software-config: adjust to new Persistent Storage implementation
We're in luck here: the Additional Software feature is enabled by default on
newly created Persistent Storage nowadays, so we don't need to explicitly
ask for it to be enabled anymore. Which is good, because the interface
to do so is not provided by the new Persistent Storage implementation yet.