Unsafe Browser: further isolate from the root filesystem
Now we are back to actually using the full chroot we set up. This
might be overkill now that we already confine with AppArmor, but it's
defense in depth and might provide inspiration for how we can do this
for all users of netnsdrop (or at least Tor Browser) in the future.

refs tails/tails#5422