WhisperBack debugging info: have df ignore filesystems of type fuse.portal
When /root/.cache/doc is mounted (which indicates that xdg-document-portal is
running as root, because users will do stuff as root that can make it happen),
Bullseye's df errors out on it.

This is fixed in sid but for Tails 5.x, let's ask df to ignore filesystems of
type fuse.portal. And IMO this is too innocuous to bother tracking the removal
of this in 6.0.

Context: https://github.com/flatpak/xdg-desktop-portal/issues/512,
https://github.com/coreutils/gnulib/pull/11.