Test suite: fix "Downloading files with the Tor Browser" scenario with Tor Browser 12 on Wayland
At least with Tor Browser 12 on Wayland, the "Save as" download dialog opens
full screen, so we never see the reload button appear. So:

 - don't wait for that button to appear
 - give the "Save as" dialog more time to appear

Unfortunately, as a consequence we lose the benefits of retry_tor in this
context. This impacts only 1 scenario so I think it's good enough and not worth
refactoring further to fix that.