Test suite: remove test whose cost/benefit has become very high
This test ensures that when Tails device that had multiple SquashFS diffs
installed before the switch to "1 single SquashFS diff" (#15281, released in
4.2) gets upgraded by our Upgrader, all obsolete SquashFS diffs are deleted.

This migration happened when users did an automatic upgrade to 4.x, with x>2.

Since then, every user of an older Tails has had to do a manual upgrade to 5.x,
which deleted those SquashFS diffs.

So let's not bother verifying that this migration path still works:
even if we broke it in a future Tails, this would not affect anyone.