initramfs: bring back including all DRM modules
As long as we had MODULES=most, the plymouth initramfs-tools hook would do this
for us. But now that we use MODULES=list, that hook skips including DRM modules
automatically, so we need to do this ourselves.

And as a consequence, we can drop our old explicit list of DRM modules to
include: they're all in kernel/drivers/gpu/drm.