Test suite: drop check about tails-version output
It's broken since the renaming of /etc/amnesia to /etc/tails.
Fixing it would require updating the test IUKs to include /etc/tails/version.
Let's not bother: the output of tails-version is not critical, we only
use it for user-facing display and our programs (e.g. Upgrader) don't make
decisions based on it (instead they use /etc/os-release).