Test suite: fix "the screen keyboard works in Tor Browser"
For some reason, since I switched to starting Tor Browser from the Overview on
this MR, this step fails: when we attempt to click the "New tab" button, instead
this selects the current URL. This only happens the second time we start Tor
Browser. To fix this, let's use the same Tor Browser instance for the 2 steps
that exercise it (DuckDuckGo and screen keyboard). Incidentally, this will make
this scenario a bit faster.