generate-languages-list: workaround the fact we can't distinguish reviewed translations from non-reviewed ones
Before Tor's migration from Transifex to Weblate, we imported only reviewed
translations, which gave us a shorter list of supported languages.
This was intentional: we wanted to make the list shorter in the Welcome
Screen, and only list languages for which we were confident we had
decent localization.

We can't do that anymore (until we do #19359), so for now, let's just
stick to a hard-coded list of the languages that had reviewed strings
in Transifex.