Do not hide the GTK3 “application menu” button in title bars
Fixes #19371.

As shown in the schema of the
`/org/gnome/desktop/wm/preferences/button-layout` setting, the `appmenu`
button is enabled by default in title bars:
<https://github.com/GNOME/gsettings-desktop-schemas/blob/gnome-3-38/schemas/org.gnome.desktop.wm.preferences.gschema.xml.in#L27-L40>

In particular, when using Wayland, this button is the only way to reach
the “application menu” in GTK3 applications.

Therefore, this button must be enabled so that users can still access
this “application menu”.